logos to grid component
