+++ app/components/templates/index.ts
@@ -3,6 +3,7 @@ export * from './errors'
 export * from './field'
 export * from './hidden-submit'
 export * from './logo'
+export * from './logo-grid'
 export * from './progress-bar'
 export * from './search-bar'
 export * from './spacer'
